CS-5855 [Contact] JS error when add contact or edit shared calendar on Firefox
Fix description: Function addCustomizedOnLoadScript of class JavascriptManager is not reload function checkView of UIContactPortlet in some case (in this issue, it's the case when user input invalid last name two times). Should use method importJavascript of class JavascriptManager instead of method addCustomizedOnLoadScript, to make sure that function checkView is imported.